backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
motor==2.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
pydantic==1.8.2
python-dotenv==0.19.0
email-validator==1.1.3
dnspython==2.1.0
